monitoring/requirements-monitoring.txt
python-dotenv == 0.21.1
pandas == 2.1.0
pytz == 2023.3
Jinja2 == 3.1.2
numpy == 1.25.2
matplotlib == 3.8.1
fastapi == 0.99.1
uvicorn == 0.23.2
requests